Refactor symbols providers for workflow language services
Use different providers for each language
