+++ server/gx-workflow-ls-format2/src/inversify.config.ts
@@ -1,10 +1,15 @@
 import { container } from "@gxwf/server-common/src/inversify.config";
-import { GxFormat2WorkflowLanguageServiceImpl } from "./languageService";
-import { WorkflowTestsLanguageServiceContainerModule } from "@gxwf/workflow-tests-language-service/src/inversify.config";
-import { GalaxyWorkflowLanguageServer, WorkflowLanguageService } from "@gxwf/server-common/src/languageTypes";
+import {
+  TYPES as COMMON_TYPES,
+  GalaxyWorkflowLanguageServer,
+  SymbolsProvider,
+  WorkflowLanguageService,
+} from "@gxwf/server-common/src/languageTypes";
 import { GalaxyWorkflowLanguageServerImpl } from "@gxwf/server-common/src/server";
-import { TYPES as COMMON_TYPES } from "@gxwf/server-common/src/languageTypes";
+import { WorkflowTestsLanguageServiceContainerModule } from "@gxwf/workflow-tests-language-service/src/inversify.config";
 import { YAMLLanguageServiceContainerModule } from "@gxwf/yaml-language-service/src/inversify.config";
+import { GxFormat2WorkflowLanguageServiceImpl } from "./languageService";
+import { GxFormat2WorkflowSymbolsProvider } from "./services/symbols";
 
 export const TYPES = {
   ...COMMON_TYPES,
@@ -22,5 +27,7 @@ container
   .bind<GalaxyWorkflowLanguageServer>(TYPES.GalaxyWorkflowLanguageServer)
   .to(GalaxyWorkflowLanguageServerImpl)
   .inSingletonScope();
+
+container.bind<SymbolsProvider>(TYPES.SymbolsProvider).to(GxFormat2WorkflowSymbolsProvider).inSingletonScope();
 
 export { container };
