+++ server/gx-workflow-ls-native/src/inversify.config.ts
@@ -1,10 +1,10 @@
 import { container } from "@gxwf/server-common/src/inversify.config";
-import { NativeWorkflowLanguageServiceImpl } from "./languageService";
-import { WorkflowTestsLanguageServiceContainerModule } from "@gxwf/workflow-tests-language-service/src/inversify.config";
-import { GalaxyWorkflowLanguageServer, WorkflowLanguageService } from "@gxwf/server-common/src/languageTypes";
+import { GalaxyWorkflowLanguageServer, TYPES, WorkflowLanguageService } from "@gxwf/server-common/src/languageTypes";
 import { GalaxyWorkflowLanguageServerImpl } from "@gxwf/server-common/src/server";
-import { TYPES } from "@gxwf/server-common/src/languageTypes";
+import { WorkflowTestsLanguageServiceContainerModule } from "@gxwf/workflow-tests-language-service/src/inversify.config";
 import { YAMLLanguageServiceContainerModule } from "@gxwf/yaml-language-service/src/inversify.config";
+import { NativeWorkflowLanguageServiceImpl } from "./languageService";
+import { NativeWorkflowSymbolsProvider } from "./services/symbols";
 
 container.load(YAMLLanguageServiceContainerModule);
 container.load(WorkflowTestsLanguageServiceContainerModule);
@@ -17,6 +17,11 @@ container
 container
   .bind<GalaxyWorkflowLanguageServer>(TYPES.GalaxyWorkflowLanguageServer)
   .to(GalaxyWorkflowLanguageServerImpl)
+  .inSingletonScope();
+
+container
+  .bind<NativeWorkflowSymbolsProvider>(TYPES.SymbolsProvider)
+  .to(NativeWorkflowSymbolsProvider)
   .inSingletonScope();
 
 export { container };
