+++ server/gx-workflow-ls-native/src/services/symbols.ts
@@ -1,0 +1,17 @@
+import { SymbolsProviderBase } from "@gxwf/server-common/src/providers/symbolsProvider";
+import { injectable } from "inversify";
+
+@injectable()
+export class NativeWorkflowSymbolsProvider extends SymbolsProviderBase {
+  constructor() {
+    super();
+    this.symbolNamesToIgnore = new Set([
+      "a_galaxy_workflow",
+      "position",
+      "uuid",
+      "errors",
+      "format-version",
+      "version",
+    ]);
+  }
+}
