+++ server/packages/workflow-tests-language-service/src/inversify.config.ts
@@ -1,4 +1,8 @@
-import { TYPES as COMMON_TYPES, WorkflowTestsLanguageService } from "@gxwf/server-common/src/languageTypes";
+import {
+  TYPES as COMMON_TYPES,
+  SymbolsProvider,
+  WorkflowTestsLanguageService,
+} from "@gxwf/server-common/src/languageTypes";
 import { ContainerModule } from "inversify";
 import { GxWorkflowTestsLanguageServiceImpl } from "./languageService";
 import { JSONSchemaService, JSONSchemaServiceImpl } from "./schema/adapter";
@@ -6,6 +10,7 @@ import { WorkflowTestsSchemaProvider, WorkflowTestsSchemaProviderImpl } from "./
 import { WorkflowTestsSchemaService, WorkflowTestsSchemaServiceImpl } from "./schema/service";
 import { WorkflowTestsCompletionService, WorkflowTestsCompletionServiceImpl } from "./services/completion";
 import { WorkflowTestsHoverService, WorkflowTestsHoverServiceImpl } from "./services/hover";
+import { WorkflowTestsSymbolsProvider } from "./services/symbols";
 import { WorkflowTestsValidationService, WorkflowTestsValidationServiceImpl } from "./services/validation";
 import { TYPES } from "./types";
 
@@ -33,4 +38,6 @@ export const WorkflowTestsLanguageServiceContainerModule = new ContainerModule((
   bind<WorkflowTestsLanguageService>(COMMON_TYPES.WorkflowTestsLanguageService)
     .to(GxWorkflowTestsLanguageServiceImpl)
     .inSingletonScope();
+
+  bind<SymbolsProvider>(TYPES.WorkflowTestsSymbolsProvider).to(WorkflowTestsSymbolsProvider).inSingletonScope();
 });
