+++ server/packages/workflow-tests-language-service/src/services/symbols.ts
@@ -1,0 +1,5 @@
+import { SymbolsProviderBase } from "@gxwf/server-common/src/providers/symbolsProvider";
+import { injectable } from "inversify";
+
+@injectable()
+export class WorkflowTestsSymbolsProvider extends SymbolsProviderBase {}
